+++ Podfile
@@ -20,7 +20,7 @@ end
 
 target 'openHABTestsSwift' do
     inherit! :search_paths
-    pod 'GDataXML-HTML', '~> 1.3.0'
+    #pod 'GDataXML-HTML', '~> 1.3.0'
     pod 'AFNetworking', '~> 2.6.0'
     pod 'Fuzi', '~> 3.1'
     #pod 'AFNetworking', '~> 3.0'
