+++ openHABTests-Bridging-Header.h
@@ -3,5 +3,5 @@
 //
 
 #import <AFNetworking/AFNetworking.h>
-#import <GDataXML_HTML/GDataXMLNode.h>
+//#import <GDataXML_HTML/GDataXMLNode.h>
